Rethink the tracker client code somewhat.
btpd now uses all tiers in parallel, so a torrent with two tiers will
essentially be treated by btpd as a torrent with two trackers to send
each event to. This is not quite what the multitrackes standard says,
but it's much easier to implement reasonable behaviour this way.

btpd is going to report the number of good trackers for a torrent
instead of the number of errors, but for now just report zero errors.