Use only IPv4 by default.
Unfortunately there are some useless IPv6 trackers that will
be used instead of their more usefull IPv4 counterparts when
IPv6 is enabled.

The -4 and -6 options are tweaked once again. If -6 is given
only IPv6 will be used. If both are given both will be used
and otherwise only IPv4 will be used.

The IPvX stuff should probably be set per torrent in the future.