Adds hook option to btpd.
Now btpd allow hook options, a hook is a script which path is passed as
argument with --hook-script=file option. The script will be executed
when torrent start leeching, stop or finished to download. And is called
with two arguments. The first one is a character which indicates the status,
as follow:

'S': Start
'T': Stop
'F': Finished

And the next argument is the torrent name.